requirements.txt

streamlit
python-docx    
PyPDF2           
scikit-learn
scipy
numpy
pandas
joblib
threadpoolctl
nltk
six
tzdata
python-dateutil==2.9.0.post0
pytz==2024
Flask==3.0.3
Werkzeug==3.1.3
Jinja2==3.1.4
itsdangerous==2.2.0
MarkupSafe==3.0.2
colorama==0.4.6
click==8.1.7
blinker==1.9.0
